@@ -1,6 +1,6 @@
 import {Injectable} from '@angular/core';
 import {Observable} from 'rxjs';
-import {User} from '../../models/user';
+import {User} from '../../model/user';
 import {HttpClient} from '@angular/common/http';
 
 @Injectable({ providedIn: 'root' })
