@@ -1,8 +1,8 @@
 import {Injectable} from '@angular/core';
 import {Observable} from 'rxjs';
-import {AuthInfo} from '../../models/auth-info';
+import {AuthInfo} from '../../model/auth-info';
 import {HttpErrorResponse} from '@angular/common/http';
-import {User} from '../../models/user';
+import {User} from '../../model/user';
 import {AuthState} from './auth.state';
 import {ConnectionIssueModalVisible} from '../../components/connection-issue-resolver/connection-issue-modal-visible';
 import {AuthApi} from './auth.api';
